pyproject.toml
[tool.poetry]
name = "boiler-plate"
version = "0.1.0"
description = ""
authors = ["Taimoor Hassan <taimoor@synares.com>"]
readme = "README.md"
packages = [{include = "boiler_plate"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
